Create Hotel Booking Analysis
This is an EDA project regarding to hotel booking analysis.We have performed EDA on the given dataset.The given dataset contains 119390 rows and 32 columns.There are mainly two hotels in the given dataset and they are City hotel and Resort hotel, etc. In this we devide work of data manipulation in three catagories which are Data collection, Data Cleaning & Manipulation, EDA. Before visualisation of any data we have done data wrangling. After that Different charts are used for data visualisation to represent the data for easily understandable format.